Make the contract test into a integrated test
